fix(ui): resolve react-refresh fast-refresh lint warnings
