+++ apps/ui/src/components/highlighted-line.tsx
@@ -1,5 +1,4 @@
-/** Mirrors the backend's <IP> placeholder substitution (ip_extract.rs). */
-export const IP_PATTERN = "\\b(?:\\d{1,3}\\.){3}\\d{1,3}\\b";
+import { IP_PATTERN } from "./ip-pattern";
 
 interface HighlightedLineProps {
   line: string;
