+++ apps/ui/src/components/ip-pattern.ts
@@ -1,0 +1,2 @@
+/** Mirrors the backend's <IP> placeholder substitution (ip_extract.rs). */
+export const IP_PATTERN = "\\b(?:\\d{1,3}\\.){3}\\d{1,3}\\b";
