+++ apps/ui/src/lib/datasource/index.ts
@@ -4,8 +4,9 @@ import type { DataSource } from "./types";
 /** The active data source for this build. */
 export const dataSource: DataSource = new HttpDataSource();
 
-export { DataSourceProvider, useDataSource } from "./context";
+export { useDataSource } from "./context";
 export { ApiError } from "./http";
+export { DataSourceProvider } from "./provider";
 export type {
   BanEvent,
   Config,
